refactor(structure): move bootstrap to app folder
